+++ src/main/java/com/robertx22/age_of_exile/config/forge/ClientConfigs.java
@@ -28,11 +28,9 @@ public class ClientConfigs {
     public boolean SHOW_LOW_ENERGY_MANA_WARNING = true;
     public boolean USE_RIGHT_CLICK_FOR_SPELL_CASTING = false;
     public boolean RENDER_SIMPLE_MOB_BAR = true;
+    public boolean RENDER_DEATH_STATISTICS = true;
 
     public int REMOVE_EMPTY_TOOLTIP_LINES_IF_MORE_THAN_X_LINES = 32;
-
-    public int LEFT_VANILLA_LIKE_BARS_Y__POS_ADJUST = 0;
-    public int RIGHT_VANILLA_LIKE_BARS_Y__POS_ADJUST = 0;
 
     public PlayerGUIs PLAYER_GUI_TYPE = PlayerGUIs.RPG;