+++ src/main/java/com/robertx22/age_of_exile/database/data/spells/entities/SimpleProjectileEntity.java
@@ -181,6 +181,7 @@ public class SimpleProjectileEntity extends PersistentProjectileEntity implement
     }
 
     public void onTick() {
+
         this.getSpellData()
             .getSpell()
             .getAttached()