+++ src/main/java/com/robertx22/age_of_exile/mmorpg/event_registers/Client.java
@@ -3,6 +3,7 @@ package com.robertx22.age_of_exile.mmorpg.event_registers;
 import com.robertx22.age_of_exile.event_hooks.ontick.OnClientTick;
 import com.robertx22.age_of_exile.event_hooks.player.OnKeyPress;
 import com.robertx22.age_of_exile.gui.overlays.bar_overlays.types.RPGGuiOverlay;
+import com.robertx22.age_of_exile.gui.overlays.death_stats.DeathStatsOverlay;
 import com.robertx22.age_of_exile.gui.overlays.mob_bar.MobBarScreen;
 import com.robertx22.age_of_exile.gui.overlays.spell_cast_bar.SpellCastBarOverlay;
 import com.robertx22.age_of_exile.gui.overlays.spell_hotbar.SpellHotbarOverlay;
@@ -20,6 +21,7 @@ public class Client {
         HudRenderCallback.EVENT.register(new MobBarScreen());
         HudRenderCallback.EVENT.register(new SpellCastBarOverlay());
         HudRenderCallback.EVENT.register(new RPGGuiOverlay());
+        HudRenderCallback.EVENT.register(new DeathStatsOverlay());
 
     }
 }