+++ src/main/java/com/robertx22/age_of_exile/vanilla_mc/packets/sync_cap/PlayerCaps.java
@@ -1,6 +1,7 @@
 package com.robertx22.age_of_exile.vanilla_mc.packets.sync_cap;
 
 import com.robertx22.age_of_exile.capability.bases.ICommonPlayerCap;
+import com.robertx22.age_of_exile.mmorpg.ModRegistry;
 import com.robertx22.age_of_exile.uncommon.datasaving.Load;
 import net.minecraft.entity.player.PlayerEntity;
 
@@ -10,6 +11,12 @@ public enum PlayerCaps {
         @Override
         public ICommonPlayerCap getCap(PlayerEntity player) {
             return Load.Unit(player);
+        }
+    },
+    DEATH_STATS {
+        @Override
+        public ICommonPlayerCap getCap(PlayerEntity player) {
+            return ModRegistry.COMPONENTS.PLAYER_DEATH_DATA.get(player);
         }
     },
     PLAYER_SKILLS {